requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-genai>=1.0.0
datasets>=2.14.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
